# docs/03-device-api.openapi.yaml
openapi: 3.0.3
info:
  title: 小卡 Device API
  version: 0.1.0
  description: |
    设备端（ESP32-S3）API 规范 v1 —— 权威文档见 `02-device-api-v1.md`。
    本文件用于导入 Postman/Apifox 调试或生成 mock server。
    Web 社区端接口（注册/登录/好友请求/绑定设备等）不在本文件范围内。
servers:
  - url: https://api-dev.example.com/v1
    description: dev 环境（联调用）
  - url: https://api.example.com/v1
    description: 生产
...
paths:
...
  /photos/{photo_id}/image:
    get:
      tags: [photos]
      summary: 下载照片（320x240 JPEG）
      parameters:
        - { name: photo_id, in: path, required: true, schema: { type: string } }
        - { name: size, in: query, required: false, schema: { type: integer, default: 320 }, description: "预留，MVP 仅一档" }
      responses:
        "200":
          description: "JPEG 字节流（需认证且通过好友可见性校验，Cache-Control: private immutable）"
          content:
            image/jpeg:
              schema: { type: string, format: binary }
        "404":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: 无权查看该照片
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/StorageUnavailable" }
...
  /photos/{photo_id}/reactions:
    post:
      tags: [feed]
      summary: 发送反应
      parameters:
        - { name: photo_id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type: { type: string, enum: [heart, thumbsup, wow] }
      responses:
        "201":
          description: 最新计数
          content:
            application/json:
              schema:
                type: object
                properties:
                  reactions:
                    type: object
                    additionalProperties: { type: integer }
                    example: { heart: 3 }
  /photos/{photo_id}/reactions/{type}:
    delete:
      tags: [feed]
      summary: 取消反应
      parameters:
        - { name: photo_id, in: path, required: true, schema: { type: string } }
        - { name: type, in: path, required: true, schema: { type: string, enum: [heart, thumbsup, wow] } }
      responses:
        "204": { description: 已取消 }
...
  /friends/claim:
    post:
      tags: [friends]
      summary: "[Phase 2 预留] 兑换离线好友票，MVP 返回 501"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ticket]
              properties:
                ticket: { type: string }
      responses:
        "501": { description: MVP 未实现 }
        "201": { description: 好友关系已建立 }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "device_token，配对成功后获得"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: PHOTO_TOO_LARGE }
            message: { type: string }
...
  responses:
    Unauthorized:
      description: "token 无效/被解绑 → 清 NVS token，回配对流程"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    StorageUnavailable:
      description: "图片存储暂不可用；请求未完成，稍后重试"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
security:
  - bearerAuth: []